==> code/00_Preprocess_Data.Rmd <==
---
title: "Appendix (Extra analyses)"
author: "Connor Quiroz"
date: "2025-02-12"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
# General data cleaning + visualization + analysis packages
library(tidyverse)
library(ggformula)
library(countrycode)
library(data.table)
library(arrow)
library(cowplot)
library(RColorBrewer)

# Map creation
library(rnaturalearth)
library(rnaturalearthdata)
library(sf)

# Load in packages needed for ARTIS
library(devtools)
library(tidytext)
# devtools::install_github("davidsjoberg/ggsankey")
library(ggsankey)
# devtools::install_github("Seafood-Globalization-Lab/exploreARTIS@v1.0.0", dependencies = TRUE)
library(exploreARTIS)

# For Packages for fishbase
# remotes::install_github("cboettig/duckdbfs", force = TRUE)
# remotes::install_github("ropensci/rfishbase")
library(duckdbfs)
library(rfishbase)
```

```{r load in functions}
# Function description
# grouping: Define which variable you want to group by
# taxa_var: input the name of the taxa variable you want to search for (e.g., "taxa_var = sciname_hs_modified" will run operations on that variable)
at_least_vars <- function(data = con_joined, grouping = "producer_iso3c", taxa_var = "sciname") {
  if (any(names(data) %in% grouping)) {
    new_data <- data %>%
      group_by(!!sym(grouping), !!sym(taxa_var)) %>%
      mutate(
        at_least_species = case_when(
          taxa_level == "species" ~ 1, TRUE ~ 0),
        at_least_genus = case_when(
          taxa_level == "species" |
          taxa_level == "genus" ~ 1, TRUE ~ 0),
        at_least_family = case_when(
          taxa_level == "species" |
          taxa_level == "genus" |
          taxa_level == "family" ~ 1,
          TRUE ~ 0
        ),
        at_least_order = case_when(
          taxa_level == "species" |
            taxa_level == "genus" |
            taxa_level == "family" |
            taxa_level == "order" ~ 1,
          TRUE ~ 0
        ),
        at_least_class = case_when(
          taxa_level == "species" |
            taxa_level == "genus" |
            taxa_level == "family" |
            taxa_level == "order" | 
            taxa_level == "class" ~ 1,
          TRUE ~ 0
        ),
        at_least_phylum = case_when(
          taxa_level == "species" |
            taxa_level == "genus" |
            taxa_level == "family" |
            taxa_level == "order" | 
            taxa_level == "class" |
            taxa_level == "phylum" ~ 1,
          TRUE ~ 0
        ),
        at_least_kingdom = case_when(
          taxa_level == "species" |
            taxa_level == "genus" |
            taxa_level == "family" |
            taxa_level == "order" | 
            taxa_level == "class" |
            taxa_level == "phylum" |
            taxa_level == "kingdom" ~ 1,
          TRUE ~ 0
        )
      )
  }
  else {
    new_data <- data %>%
      group_by(!!sym(taxa_var)) %>%
      mutate(
        at_least_species = case_when(taxa_level == "species" ~ 1, TRUE ~ 0),
        at_least_genus = case_when(taxa_level == "species" |
                                     taxa_level == "genus" ~ 1, TRUE ~ 0),
        at_least_family = case_when(
          taxa_level == "species" |
            taxa_level == "genus" | taxa_level == "family" ~ 1,
          TRUE ~ 0
        ),
        at_least_order = case_when(
          taxa_level == "species" |
            taxa_level == "genus" |
            taxa_level == "family" | taxa_level == "order" ~ 1,
          TRUE ~ 0
        ),
        at_least_class = case_when(
          taxa_level == "species" |
            taxa_level == "genus" |
            taxa_level == "family" |
            taxa_level == "order" | 
            taxa_level == "class" ~ 1,
          TRUE ~ 0
        )
      )
  }
  
  # Copy input data to use throughout loop
  data_1 <- data
  # Calculate proportions for which a country reports to X taxa level
  for (i in c("species", "genus", "family", "order", "class", "phylum", "kingdom")) {
  con_prop_group <- new_data %>%
    group_by(!!sym(grouping)) %>%
    summarize(!!paste0("prop_at_least_", i) := sum(!!sym(paste0("at_least_",i))) / n()) %>%
    ungroup()
  
  # Join prop of each phylogenetic group to already joined consumption dataset
  data_1 <- data_1 %>%
      left_join(con_prop_group, by = grouping)

  }
  return(data_1)
}

taxa_level_vars <- function(data, taxa_var = "sciname",ungroup = FALSE) {
  if (ungroup == FALSE) {
    data %>%
      mutate(
        taxa_level = case_when(
          !!sym(taxa_var) == kingdom ~ "kingdom",
          !!sym(taxa_var) == phylum ~ "phylum",
          !!sym(taxa_var) == superclass ~ "superclass",
          !!sym(taxa_var) == class ~ "class",
          !!sym(taxa_var) == order ~ "order",
          !!sym(taxa_var) == family ~ "family",
          !!sym(taxa_var) == subfamily ~ "subfamily",
          !!sym(taxa_var) == genus ~ "genus",
          str_detect(!!sym(taxa_var), pattern = "\ ") ~ "species"
        )
      )
  }
  
  else {
    data %>%
      ungroup() %>%
      distinct(sciname) %>%
      left_join(sciname, by = "sciname") %>%
      mutate(
        taxa_level = case_when(
          !!sym(taxa_var) == kingdom ~ "kingdom",
          !!sym(taxa_var) == phylum ~ "phylum",
          !!sym(taxa_var) == superclass ~ "superclass",
          !!sym(taxa_var) == class ~ "class",
          !!sym(taxa_var) == order ~ "order",
          !!sym(taxa_var) == family ~ "family",
          !!sym(taxa_var) == subfamily ~ "subfamily",
          !!sym(taxa_var) == genus ~ "genus",
          str_detect(!!sym(taxa_var), pattern = "\ ") ~ "species"
        )
      )
  }
  
  
}

# fishbase::country()

# Another way to get species distributions in certain countries (need to see how many countries are represented in this vs FAO, which doesn't directly report countries, but needs to be joined on eez data).
taxa_presence_or_absence <- function(data = "fishbase", obtain_stats = FALSE) {
  
  # Determine whether to use FB or SLB data
  if (data == "fishbase") {
    artis_fishbase_locations <- fb_presences_country
    species_codes <- fb_species_codes
  }
  else {
    artis_fishbase_locations <- slb_presences_country
    species_codes <- slb_species_codes
  }
  
  if (obtain_stats == TRUE) {
    # faoareas() and country() have equal coverage - 1075 species from ARTIS' 1326 species
    # 229 countries when dealing with country() data, but how many covered countries when dealing with faoareas() data?
    num_countries <- artis_fishbase_locations %>%
      filter(Status %in% c("native", "Native", "introduced", "endemic") & !is.na(country)) %>%
      mutate(eez_name = countrycode(country, origin = 'country.name', destination = 'iso3c')) %>%
      distinct(eez_name) %>%
      summarize(length = length(eez_name)) %>%
      pull(length)
    print(paste0("Number of countries joined to ", data, ": ", as.character(num_countries)))
    
    # Data coverage:
    num_species_artis <- sum(str_detect(
      unique(scinames_excluding_tradeflows$sciname_joinable_to_fishbase), " "), na.rm = TRUE)
    num_species_fishbase <- artis_fishbase_locations %>%
      summarize(unique_species = length(unique(Species))) %>%
      pull(unique_species)
    
    species_coverage_pct <- round((num_species_fishbase / num_species_artis), 2) * 100
    
    # 81% % coverage of all ARTIS species in fishbase - also try sealifebase?
    print(paste0("Number of species in ", sym(data), ": ", num_species_fishbase))
    print(paste0("Number of species in ARTIS", ": ", num_species_artis))
    print(paste0("Percent of ARTIS species that ", sym(data), " has: ", as.character(species_coverage_pct),"%"))
  }
  else {
    # Which countries species are found in (using country() distribution data)
    fishbase_country_presences <- artis_fishbase_locations %>%
      filter(Status %in% c("native", "Native", "introduced", "endemic")) %>%
      mutate(eez_name = countrycode(country, origin = 'country.name', destination = 'iso3c')) %>%
      group_by(ScientificName) %>%
      summarize(countries_found = paste0(unique(eez_name), collapse = " ")) %>%
      rename(sciname_joinable_to_fishbase = "ScientificName") %>%
      mutate(source = data)
    
    consumption_countries_joined <- left_join(scinames_excluding_tradeflows, fishbase_country_presences, by = "sciname_joinable_to_fishbase")
    
    # # See which entries were successfully joined
    # consumption_countries_joined <- consumption_countries_joined %>%
    #   mutate(
    #     joined = case_when(!is.na(eez_iso3c) & !is.na(countries_found) ~ "both",
    #                        !is.na(eez_iso3c) ~ "only in artis",
    #                        TRUE ~ paste0("only in ", data)),
    #     source = case_when(joined == "both" ~ data,
    #                        joined == paste0("only in ", data) ~ data,
    #                        TRUE ~ NA))

    # Determine whether a species is present or absent in an EEZ
    consumption_countries_joined <- consumption_countries_joined %>%
      ungroup() %>%
      mutate(presence_or_absence = case_when(
        str_detect(countries_found, eez_iso3c) ~ 1, !str_detect(countries_found, eez_iso3c) ~ 0,
        TRUE ~ NA
      ))
    
    return(consumption_countries_joined)
  }
  
  
}
```

```{r consuption data reading}
# Read in consumption data
consumption_all_years <- read_parquet("../data/consumption/example_consumption_eez_2024_12_06.parquet")

# Filter working data to just 2019
consumption <- consumption_all_years %>%
  filter(year == 2019)

# Read in taxonomic data
sciname <- read.csv("../data/consumption/sciname.csv")

# Get unique taxa names by eez (excluding trade flows)
scinames_excluding_tradeflows <- consumption_all_years %>%
  ungroup() %>%
  distinct(eez_iso3c, sciname_hs_modified) %>%
  left_join(sciname, by = c("sciname_hs_modified" = "sciname")) %>%
  rename(sciname = "sciname_hs_modified") %>%
  taxa_level_vars() %>%
  rename(sciname_hs_modified = "sciname") %>%
  filter(!is.na(sciname_hs_modified)) %>%
  ungroup()

# Write to parquet
write_parquet(scinames_excluding_tradeflows, "../output/scinames_excluding_tradeflows.parquet")
```

```{r get fb/slb data}
if (!file.exists("../data/fb_slb_data/")) {
  
  # Create new directory for FB/SLB data
  dir.create("../data/fb_slb_data")
  
  # Obtain FB / SLB country() data
  fb_presences_country <- country(server = "fishbase")
  slb_presences_country <- country(server = "sealifebase")
  
  fb_species_codes <- load_taxa(server = "fishbase") %>%
    rename(ScientificName = "Species") %>%
    mutate(Species = sapply(str_split(ScientificName, " "), `[`, 2)) %>%
    relocate(Species, .before = Genus)

  slb_species_codes <- load_taxa(server = "sealifebase") %>%
    rename(ScientificName = "Species") %>%
    mutate(Species = sapply(str_split(ScientificName, " "), `[`, 2)) %>%
    relocate(Species, .before = Genus)
  
  # Join FB / SLB country() to species codes
  fb_presences_country <- fb_presences_country %>%
    left_join(fb_species_codes, by = "SpecCode") %>%
    rename(SubFamily = "Subfamily") %>%
    mutate(Phylum = NA_character_, Kingdom = NA_character_) %>% # Create new 
    relocate(c(ScientificName, Species, Genus, SubFamily, Family, Order, Class, SuperClass, Phylum, Kingdom), .after = SpecCode) %>% # Move scinames to the front of dataset
    mutate(across(c(ScientificName, Species, Genus, SubFamily, Family, Order, Class, SuperClass, Phylum, Kingdom), ~ str_to_lower(.))) # convert scinames to lowercase
  
  # Same thing but for sealifebase
  slb_presences_country <- slb_presences_country %>%
    left_join(slb_species_codes, by = "SpecCode") %>%
    rename(SubFamily = "Subfamily") %>%
    mutate(SuperClass = NA_character_) %>%
    relocate(c(ScientificName, Species, Genus, SubFamily, Family, Order, Class, SuperClass, Phylum, Kingdom), .after = SpecCode) %>%
    mutate(across(c(ScientificName, Species, Genus, SubFamily, Family, Order, Class, SuperClass, Phylum, Kingdom), ~ str_to_lower(.))) 
  
  # Write files to .parquet
  write_parquet(fb_presences_country, "../data/fb_slb_data/fb_presences_country.parquet")
  write_parquet(slb_presences_country, "../data/fb_slb_data/slb_presences_country.parquet")
  write_parquet(slb_species_codes, "../data/fb_slb_data/fb_species_codes.parquet")
  write_parquet(slb_species_codes, "../data/fb_slb_data/slb_species_codes.parquet")
}
```

```{r add in fishbase / sealifebase data connections}
# Read in consumption data
scinames_excluding_tradeflows <- read_parquet("../output/scinames_excluding_tradeflows.parquet")

# Read in fishbase/sealifebase data
fb_presences_country <- read_parquet("../data/fb_slb_data/fb_presences_country.parquet")
slb_presences_country <- read_parquet("../data/fb_slb_data/slb_presences_country.parquet")

fb_species_codes <- read_parquet("../data/fb_slb_data/fb_species_codes.parquet")
slb_species_codes <- read_parquet("../data/fb_slb_data/slb_species_codes.parquet")

scinames_excluding_tradeflows <- scinames_excluding_tradeflows %>%
  mutate(sciname_joinable_to_fishbase = sciname_hs_modified)

# Modify sciname casing so that it can be joined to fishbase
# scinames_excluding_tradeflows <- scinames_excluding_tradeflows %>%
#   mutate(sciname_joinable_to_fishbase = str_to_title(word(sciname_hs_modified, 1)), 
#          sciname_joinable_to_fishbase2 = word(sciname_hs_modified, 2), 
#          sciname_joinable_to_fishbase = ifelse(
#         is.na(sciname_joinable_to_fishbase2),
#         sciname_joinable_to_fishbase,
#         paste(sciname_joinable_to_fishbase, sciname_joinable_to_fishbase2))) %>%
#   select(-sciname_joinable_to_fishbase2) %>%
#   relocate(sciname_joinable_to_fishbase, .after = sciname_hs_modified)

# ***
# MAKING SURE ARTIS SCINAMES ALIGN WITH FISHBASE

# # of ARTIS unique ID's (includes higher taxonomic levels that aren't only species) - 1771 total IDs
length(unique(scinames_excluding_tradeflows$sciname_joinable_to_fishbase))

# # of ARTIS species (1421 species)
scinames_excluding_tradeflows %>%
  ungroup() %>%
  filter(str_detect(sciname_hs_modified, " ")) %>%
  distinct(sciname_hs_modified) %>%
  pull(sciname_hs_modified) %>%
  length()

# Total number of joined species succesfully from ARTIS
# length(unique(artis_fishbase_locations$Species))
# sum(str_detect(unique(artis_fishbase_locations$Species), " "))

# ***


# METHOD 2: rfishbase::country()

# Obtain stats on fishbase/sealifebase data
taxa_presence_or_absence(data = "fishbase", obtain_stats = TRUE)
taxa_presence_or_absence(data = "sealifebase", obtain_stats = TRUE)

# Create presence/absence data
fishbase_presences_country <- taxa_presence_or_absence(data = "fishbase")
sealifebase_presences_country <- taxa_presence_or_absence(data = "sealifebase")

# Number of rows in ARTIS data (fishbase + sealifebase presence data)
nrow(fishbase_presences_country)
nrow(sealifebase_presences_country)

# Combine fishbase/sealifebase data
merged_presence_absence_data_country <- bind_rows(fishbase_presences_country[-which(sealifebase_presences_country$source == "sealifebase"),], sealifebase_presences_country %>% filter(source == "sealifebase"))

# Add true species column
merged_presence_absence_data_country <- merged_presence_absence_data_country %>%
  mutate(species = case_when(str_detect(sciname_hs_modified, " ") ~ sapply(str_split(sciname_hs_modified, " "), `[`, 2),
                             TRUE ~ NA))



# Determine if lower resolution taxa are in an eez (e.g., if a species for a given country is found there that is also a lower resolution taxa like actinoptergyii, then all actinoptergyii or higher will be found in that country) -

# List of ARTIS species that aren't matched to either fishbase or sealifebase
# Create a dataset to identify whether a species comes from ARTIS or slb/fb
species_determination <- full_join(
  bind_rows(fb_presences_country %>% 
              mutate(source = "fishbase"), 
            slb_presences_country %>% 
              mutate(source = "sealifebase")) %>%
    mutate(ScientificName = tolower(ScientificName)) %>%
    select(ScientificName) %>%
    distinct() %>% 
    mutate(fb_slb = 1), 
  sciname %>% 
    select(sciname, common_name) %>%
    filter(str_detect(sciname, " ")) %>% 
    distinct() %>% 
    mutate(artis = 1), 
  by = c("ScientificName" = "sciname")) 

# Species in ARTIS but not in FB/SLB
artis_species_not_in_fbslb <- species_determination %>%
  filter(is.na(fb_slb) == TRUE, artis == 1) %>%
  pull(ScientificName)

# Species in FB/SLB but not ARTIS
fbslb_species_not_in_artis <- species_determination %>%
  filter(fb_slb == 1, is.na(artis) == TRUE) %>%
  pull(ScientificName)
  
pa_only_artis <- merged_presence_absence_data_country %>%
  mutate(presence_or_absence = case_when(sciname_joinable_to_fishbase %in% fbslb_species_not_in_artis ~ NA, TRUE ~ presence_or_absence))

presence_absence_data_linked <- data.frame()
for (i in unique(pa_only_artis$eez_iso3c)) {
  # Group data by eez country
  sliced_data <- pa_only_artis %>%
  filter(eez_iso3c == sym(i))

  # Create updated presence on sliced data
  sliced_data_for_merging <- sliced_data %>%
  mutate(unique_vals = str_replace_all(paste0(filter(., presence_or_absence == 1) %>% select(sciname_hs_modified, species, genus, subfamily, family, order, class, superclass, phylum, kingdom) %>%
  unlist() %>%
  unique(), collapse = " "), "[A-Z]", ""), presence_or_absence_all_taxa = case_when(sciname_hs_modified %in% fbslb_species_not_in_artis ~ NA, 
presence_or_absence == 1 ~ 1,
str_detect(unique_vals, sciname_hs_modified) ~ 1,
                                                                                    TRUE ~ 0))

  # Combine rows of each country
  presence_absence_data_linked <- bind_rows(presence_absence_data_linked, sliced_data_for_merging)

}

# Count Number of presences/absences in filled in presence data
presence_absence_data_linked <- presence_absence_data_linked %>%
  relocate(presence_or_absence_all_taxa, .after = presence_or_absence)

# Count number of total presences / absences across ALL taxa levels
presence_absence_data_linked %>%
  count(presence_or_absence_all_taxa)

# List of **linked** countries proportions of stock reported as absent
linked_props <- presence_absence_data_linked %>%
  group_by(eez_iso3c) %>%
  distinct(sciname_hs_modified, .keep_all = TRUE) %>%
  select(sciname_hs_modified, presence_or_absence_all_taxa) %>%
  count(presence_or_absence_all_taxa) %>%
  mutate(presence_or_absence_all_taxa = as.character(presence_or_absence_all_taxa)) %>%
  pivot_wider(values_from = n, names_from = presence_or_absence_all_taxa, names_glue = "quantity_{presence_or_absence_all_taxa}") %>%
  mutate(prop_missing =  quantity_0 / (quantity_0 + quantity_1))

# See which countries have only absences and no presences
presence_absence_data_linked %>%
  group_by(eez_iso3c) %>%
  distinct(sciname_hs_modified, .keep_all = TRUE) %>%
  select(sciname_hs_modified, presence_or_absence_all_taxa) %>%
  count(presence_or_absence_all_taxa) %>%
  mutate(presence_or_absence_all_taxa = as.character(presence_or_absence_all_taxa)) %>%
  pivot_wider(values_from = n, names_from = presence_or_absence_all_taxa, names_glue = "quantity_{presence_or_absence_all_taxa}") %>%
  mutate(prop_missing =  quantity_0 / (quantity_0 + quantity_1)) %>%
  filter(is.na(prop_missing) & is.na(quantity_1)) %>%
  rename(number_absences = "quantity_0", number_presences = "quantity_1")

# Get species that aren't apart of artis, but are only included in fishbase/sealife base (extra species locations we need to )
pa_excluding_artis <- merged_presence_absence_data_country %>%
  filter(str_detect(sciname_joinable_to_fishbase, " ") & !(joined %in% c("only in artis", "both")))

# See which species aren't matched
# 44 species that aren't mathed to either fishbase or sealifebase
pa_only_artis %>%
  filter(str_detect(sciname_joinable_to_fishbase, " ") & is.na(presence_or_absence)) %>%
  distinct(sciname_joinable_to_fishbase)

# Sum: 93447 - remove species from ARTIS that aren't in fishbase/sealifebase
sum(pa_excluding_artis %>%
  distinct(sciname_joinable_to_fishbase) %>%
  summarize(n = n()) %>%
  pull(n),

pa_only_artis %>%
  filter(str_detect(sciname_joinable_to_fishbase, " ") & !is.na(presence_or_absence)) %>%
  distinct(sciname_joinable_to_fishbase) %>%
  drop_na() %>%
  summarize(n = n()) %>%
  pull(n))

# Sum: 93447
sum(fb_presences_country %>%
      filter(Status %in% c("native", "Native", "introduced", "endemic") & !is.na(country)) %>%
  distinct(Species) %>%
    drop_na() %>%
  summarize(n = n()) %>%
  pull(n),
slb_presences_country %>%
  filter(Status %in% c("native", "Native", "introduced", "endemic") & !is.na(country)) %>%
  distinct(Species) %>%
  drop_na() %>%
  summarize(n = n()) %>%
  pull(n))

# Determine different commercial usages (need to only include commercial fishes for our analysis)
fb_presences_country %>%
  distinct(Importance)

# Get only non-artis species that are of commercial interest
fb_slb_commercial_interests <-  anti_join(
  bind_rows(fb_presences_country %>%
              mutate(source = "fishbase"), 
            slb_presences_country %>% 
              mutate(source = "sealifebase")), 
  sciname %>% 
    filter(str_detect(sciname, " ")) %>% 
    distinct(), 
  by = c("ScientificName" = "sciname")
  ) %>%
  filter((Status %in% c("native", "Native", "introduced", "endemic") & !is.na(country)) & (Importance %in% c("minor commercial", "highly commercial", "commercial", "subsistence fisheries", "of potential interest"))) # Get only commercial species

# Convert fb/slb commercial marine country names to iso3c codes + filter out countries NOT in ARTIS
fb_slb_commercial_interests <- fb_slb_commercial_interests %>%
  select(country, ScientificName, Species, Genus, SubFamily, Family, Order, Class, SuperClass, Phylum, Kingdom) %>%
    mutate(country = countrycode(country, origin = 'country.name', destination = 'iso3c')) %>%
    filter(country %in% unique(pa_only_artis$eez_iso3c)) %>%
    group_by(ScientificName) %>%
  mutate(countries_found = paste0(country, collapse = " ")) %>%
  distinct(ScientificName, .keep_all = TRUE) %>%
  mutate(presence_or_absence = 1) %>% # These species are found in the countries, so we assign a 1 ... will be added into artis available taxa for a country in subsequent code lines
  rename(sciname_hs_modified = "ScientificName", species = "Species", genus = "Genus", subfamily = "SubFamily", family = "Family", order = "Order", class = "Class", superclass = "SuperClass", phylum = "Phylum", kingdom = "Kingdom", eez_iso3c = "country")

# FB/SLB does not have superclass/phylum/kingdom, so we will first (1) match with artis to obtain superclass, phylum, and kingdom, and fill in the rest with rgbif

# Get class names for fb_slb data (will combine with artis_scinames to get phylum, kingdom, which fishbase does not have)
class_taxa_names <- fb_slb_commercial_interests %>%
  ungroup() %>%
  distinct(class) %>%
  pull(class)

# Create an empty list
superclass_phylum_kingdom <- data.frame(class = class_taxa_names, superclass = NA_character_, phylum = NA_character_, kingdom = NA_character_)
for (i in class_taxa_names) {
  matched_scinames <- sciname %>%
    filter(class == i) %>%
    distinct(superclass, phylum, kingdom) %>%
    select(superclass, phylum, kingdom)
  
  # Check to see if dataframe is empty or not
  if (dim(matched_scinames)[1] == 0) {
    matched_scinames <- data.frame(superclass = NA_character_,
                                            phylum = NA_character_,
                                            kingdom = NA_character_)
  }
  k <- which(class_taxa_names == i) # Currentl index in class list
  # Add entries into a list
  superclass_phylum_kingdom$superclass[k] <- matched_scinames$superclass
  superclass_phylum_kingdom$phylum[k] <- matched_scinames$phylum
  superclass_phylum_kingdom$kingdom[k] <- matched_scinames$kingdom
}

# Exclude rows that already completely filled (rows with no NAs)
taxa_to_be_filled <- superclass_phylum_kingdom %>%
  filter(is.na(superclass) |
         is.na(phylum) |
         is.na(kingdom))


# New script to refer to to obtain Global Biodiversity Information Facility (rgbif) taxa data - refer to "accepted" column to filter for target species that we will be getting the missing phylum / kingdom names

# taxa_to_be_filled is fed through this script to fill in phylum/kingdom names (takes awhile hence why another script was created)
rmarkdown::render("../code/QuirozConnor_SGL_Obtain_gbif_Data.Rmd")
taxa_to_be_filled_gbif <- read_parquet("../data/rgbif_data/taxa_to_be_filled_gbif.parquet")

# Combine rows fb/slb unlinked names that didn't need filling in + filled in phylum/kingdom names obtained via gbif
filled_fb_slb_taxa <- bind_rows(superclass_phylum_kingdom %>%
  filter(!is.na(superclass) &
         !is.na(phylum) &
         !is.na(kingdom)), taxa_to_be_filled_gbif)


# Fill in the rest of the phylum/kingdom names with gbif
fb_slb_commercial_interests <- left_join(fb_slb_commercial_interests, filled_fb_slb_taxa, by = "class") %>%
mutate(
  superclass = coalesce(superclass.x, superclass.y), # If superclass.x is NA, take superclass.y
  phylum = coalesce(phylum.x, phylum.y),  # Similarly for phylum
  kingdom = coalesce(kingdom.x, kingdom.y)) %>% # Similarly for kingdom
select(-phylum.x, -phylum.y, -kingdom.x, -kingdom.y, -superclass.x, -superclass.y)

# Re-determine if lower resolution taxa are present with both ARTIS + non-ARTIS species
only_artis_1 <- merged_presence_absence_data_country

# Combine only artis species with fb/slb commercial species NOT in artis (used to increase amount of presences in data via country function)
pa_only_artis <- bind_rows(only_artis_1, fb_slb_commercial_interests %>% mutate(source = "unlinked")) %>%
  mutate(presence_or_absence = case_when(sciname_joinable_to_fishbase %in% fbslb_species_not_in_artis ~ NA, TRUE ~ presence_or_absence))

# Create a new emptt dataframe
presence_absence_data_linked_unlinked <- data.frame()

for (i in unique(pa_only_artis$eez_iso3c)) {
  # Group data by eez country
  sliced_data <- pa_only_artis %>%
  filter(eez_iso3c == i)

  # Create updated presence on sliced data
  sliced_data_for_merging <- sliced_data %>%
  mutate(unique_vals = str_replace_all(paste0(filter(., presence_or_absence == 1) %>% select(sciname_hs_modified, species, genus, subfamily, family, order, class, superclass, phylum, kingdom) %>%
  unlist() %>%
  unique(), collapse = " "), "[A-Z]", ""), 
  presence_or_absence_all_taxa = case_when(str_detect(unique_vals, sciname_hs_modified) ~ 1,
sciname_hs_modified %in% fbslb_species_not_in_artis ~ NA,
TRUE ~ 0))

  # Combine rows of each country
  presence_absence_data_linked_unlinked <- bind_rows(presence_absence_data_linked_unlinked, sliced_data_for_merging)

}

# Count Number of presences/absences in filled in presence data
presence_absence_data_linked_unlinked <- presence_absence_data_linked_unlinked %>%
  relocate(presence_or_absence_all_taxa, .after = presence_or_absence)

# Code that displays ARTIS taxa are present/absent
presence_absence_data_linked_unlinked %>%
  filter(source != "unlinked" | is.na(source)) %>%
  count(presence_or_absence_all_taxa)

# See which countries are absent
linked_unlinked_props <- presence_absence_data_linked_unlinked %>%
  group_by(eez_iso3c) %>%
  distinct(sciname_hs_modified, .keep_all = TRUE) %>%
  select(sciname_hs_modified, presence_or_absence_all_taxa) %>%
  count(presence_or_absence_all_taxa) %>%
  mutate(presence_or_absence_all_taxa = as.character(presence_or_absence_all_taxa)) %>%
  pivot_wider(values_from = n, names_from = presence_or_absence_all_taxa, names_glue = "quantity_{presence_or_absence_all_taxa}") %>%
  mutate(prop_missing =  quantity_0 / (quantity_0 + quantity_1)) %>%
  rename(number_absences = "quantity_0", number_presences = "quantity_1")

# Only countries with missing data
linked_unlinked_props %>%
  filter(is.na(prop_missing) & is.na(number_presences))

presence_absence_data_linked_unlinked %>%
  select(-c(isscaap, sciname_joinable_to_fishbase, countries_found, presence_or_absence, unique_vals)) %>%
  relocate(species, .before = genus) %>%
  filter(is.na(source)) %>%
  ungroup() %>%
  filter(str_detect(sciname_hs_modified, " ")) %>%
  distinct(sciname_hs_modified) %>%
  arrange(sciname_hs_modified)

country

full_join(sciname,
          bind_rows(fb_presences_country, slb_presences_country) %>% 
  distinct(ScientificName) %>% 
    mutate(present = 1),
  by = c("sciname" = "ScientificName")) %>%
  filter(is.na(present)) %>%
  filter(str_detect(sciname, " ")) %>%
  distinct(sciname) %>%
  arrange(sciname)
```
