==> code/00_Preprocess_Data_Revision.Rmd <==
---
title: "Appendix (Extra analyses)"
author: "Connor Quiroz"
date: "2025-02-12"
output: html_document
---

> Only species that isn't up to date with sealifebase: Bathyraja irrasa. The data previously obtained from sealifebase says it was from "Heard McDon Is.", even though sealifebase now says it is from Kerguelen island.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
# General data cleaning + visualization + analysis packages
library(tidyverse)
library(ggformula)
library(countrycode)
library(data.table)
library(arrow)
library(cowplot)
library(RColorBrewer)

# Map creation
library(rnaturalearth)
library(rnaturalearthdata)
library(sf)

# Load in packages needed for ARTIS
library(devtools)
library(tidytext)
# devtools::install_github("davidsjoberg/ggsankey")
library(ggsankey)
# devtools::install_github("Seafood-Globalization-Lab/exploreARTIS@v1.0.0", dependencies = TRUE)
library(exploreARTIS)

# For Packages for fishbase
# remotes::install_github("cboettig/duckdbfs", force = TRUE)
# remotes::install_github("ropensci/rfishbase")
library(duckdbfs)
library(rfishbase)
```

```{r load in functions}
taxa_level_vars <- function(data, taxa_var = "sciname",ungroup = FALSE) {
  if (ungroup == FALSE) {
    data %>%
      mutate(
        taxa_level = case_when(
          !!sym(taxa_var) == kingdom ~ "kingdom",
          !!sym(taxa_var) == phylum ~ "phylum",
          !!sym(taxa_var) == superclass ~ "superclass",
          !!sym(taxa_var) == class ~ "class",
          !!sym(taxa_var) == order ~ "order",
          !!sym(taxa_var) == family ~ "family",
          !!sym(taxa_var) == subfamily ~ "subfamily",
          !!sym(taxa_var) == genus ~ "genus",
          str_detect(!!sym(taxa_var), pattern = "\ ") ~ "species"
        )
      )
  }
  
  else {
    data %>%
      ungroup() %>%
      distinct(sciname) %>%
      left_join(sciname, by = "sciname") %>%
      mutate(
        taxa_level = case_when(
          !!sym(taxa_var) == kingdom ~ "kingdom",
          !!sym(taxa_var) == phylum ~ "phylum",
          !!sym(taxa_var) == superclass ~ "superclass",
          !!sym(taxa_var) == class ~ "class",
          !!sym(taxa_var) == order ~ "order",
          !!sym(taxa_var) == family ~ "family",
          !!sym(taxa_var) == subfamily ~ "subfamily",
          !!sym(taxa_var) == genus ~ "genus",
          str_detect(!!sym(taxa_var), pattern = "\ ") ~ "species"
        )
      )
  }
  
  
}

# fishbase::country()

# Another way to get species distributions in certain countries (need to see how many countries are represented in this vs FAO, which doesn't directly report countries, but needs to be joined on eez data).
taxa_presence_or_absence <- function(data = "fishbase", obtain_stats = FALSE) {
  
  # Determine which dataset to use
  if (data == "fishbase") {
    artis_fishbase_locations <- fb_presences_country
  } else if (data == "sealifebase") {
    artis_fishbase_locations <- slb_presences_country
  } else {
    stop("Data parameter must be 'fishbase' or 'sealifebase'")
  }
  
  if (obtain_stats == TRUE) {
    # Calculate statistics
    valid_locations <- artis_fishbase_locations %>%
      filter(Status %in% c("native", "Native", "introduced", "endemic", "Endemic") & 
             !is.na(country)) %>%
      mutate(eez_name = case_when(
        country == "Micronesia" ~ "FSM",
        TRUE ~ countrycode(country, 
                           origin = 'country.name',
                           destination = 'iso3c', 
                           warn = FALSE))) %>%
      filter(!is.na(eez_name))
    
    num_countries <- valid_locations %>%
      distinct(eez_name) %>%
      nrow()
    
    print(paste0("Number of countries joined to ", data, ": ", num_countries))
    
    # Data coverage calculations
    if (exists("scinames_excluding_tradeflows")) {
      num_species_artis <- scinames_excluding_tradeflows %>%
        filter(str_count(sciname_joinable_to_fishbase, "\\S+") >= 2) %>%
        distinct(sciname_joinable_to_fishbase) %>%
        nrow()
      
      num_species_fishbase <- valid_locations %>%
        distinct(ScientificName) %>%
        nrow()
      
      species_coverage_pct <- round((num_species_fishbase / num_species_artis) * 100, 2)
      
      print(paste0("Number of species in ", data, ": ", num_species_fishbase))
      print(paste0("Number of species in ARTIS: ", num_species_artis))
      print(paste0("Percent of ARTIS species that ", data, " has: ", species_coverage_pct, "%"))
    }
    
  } else {
    # Create presence/absence data
    
    # FIXED: Simplified country presence logic
    fishbase_country_presences <- artis_fishbase_locations %>%
      filter(Status %in% c("native", "Native", "introduced", "endemic", "Endemic")) %>%
      mutate(eez_name = case_when( # Correct specific countries since countrycode won't
        country == "Micronesia" ~ "FSM", # Fix Micronesia
        country == "Dominican Rp" ~ "DOM", # fix Dominical Republic
        country == "Kerguelen Is." ~ "FRA", # Territory fix
        country == "Heard McDon Is." ~ "AUS", # Territory fix
        country == "Hawaii" ~ "USA", # Territory fix
        TRUE ~ countrycode(country, 
                           origin = 'country.name',
                           destination = 'iso3c', 
                           warn = FALSE))) %>%
      filter(!is.na(eez_name)) %>%
      group_by(ScientificName) %>%
      summarize(
        countries_found = paste(unique(eez_name), collapse = " "),
         .groups = 'drop'
      ) %>%
      rename(sciname_joinable_to_fishbase = ScientificName) %>%
      mutate(source = data) %>% 
  left_join(unaligned_names[,c("fb_slb_scientific_name", "new_sciname")], by = c("sciname_joinable_to_fishbase" = "fb_slb_scientific_name")) %>% # Reassign some fb/slb names to match the updated names that ARTIS uses
  mutate(sciname_joinable_to_fishbase = case_when(!is.na(new_sciname) ~ new_sciname, 
                                    TRUE ~ sciname_joinable_to_fishbase))
    
    # Join with consumption data
    if (exists("scinames_excluding_tradeflows")) {
      consumption_countries_joined <- scinames_excluding_tradeflows %>%
        left_join(fishbase_country_presences, by = "sciname_joinable_to_fishbase")
      
      # FIXED: Proper presence/absence determination
      consumption_countries_joined <- consumption_countries_joined %>%
        mutate(
          presence_or_absence = case_when(
            is.na(countries_found) ~ NA_real_,
            str_detect(countries_found, paste0("\\b", eez_iso3c, "\\b")) ~ 1,
            TRUE ~ 0
          )
        ) %>% mutate(
        joined = case_when(!is.na(eez_iso3c) & !is.na(countries_found) ~ "both",
                           !is.na(eez_iso3c) ~ "only in artis",
                           TRUE ~ paste0("only in ", data)),
        source = case_when(joined == "both" ~ data,
                           joined == paste0("only in ", data) ~ data,
                           TRUE ~ NA))
      
      return(consumption_countries_joined)
    } else {
      warning("scinames_excluding_tradeflows not found")
      return(fishbase_country_presences)
    }
  }
}

# Extract countries that are present
get_country_taxa_data <- function(data) {
  # Genus
  genus_country_lookup <- pa_only_artis %>%
    filter(presence_or_absence == 1, !is.na(genus)) %>%
    mutate(country = strsplit(countries_found, " ")) %>%
    unnest(country) %>%
    distinct(genus, country) %>%
    group_by(genus) %>%
    summarize(countries_for_genus = paste(sort(unique(country)), collapse = " "),
              .groups = "drop") %>%
    {
      setNames(.$countries_for_genus, .$genus)
    } %>%
    data.table(
      taxa_name = names(.),
      countries_found = .,
      taxa_level = "genus"
    )
  
  # Subfamily
  subfamily_country_lookup <- pa_only_artis %>%
    filter(presence_or_absence == 1, !is.na(subfamily)) %>%
    mutate(country = strsplit(countries_found, " ")) %>%
    unnest(country) %>%
    distinct(subfamily, country) %>%
    group_by(subfamily) %>%
    summarize(countries_for_genus = paste(sort(unique(country)), collapse = " "),
              .groups = "drop") %>%
    {
      setNames(.$countries_for_genus, .$subfamily)
    } %>%
    data.table(
      taxa_name = names(.),
      countries_found = .,
      taxa_level = "subfamily"
    )
  
  # Family
  family_country_lookup <- pa_only_artis %>%
    filter(presence_or_absence == 1, !is.na(family)) %>%
    mutate(country = strsplit(countries_found, " ")) %>%
    unnest(country) %>%
    distinct(family, country) %>%
    group_by(family) %>%
    summarize(countries_for_genus = paste(sort(unique(country)), collapse = " "),
              .groups = "drop") %>%
    {
      setNames(.$countries_for_genus, .$family)
    } %>%
    data.table(
      taxa_name = names(.),
      countries_found = .,
      taxa_level = "family"
    )
  
  # Order
  order_country_lookup <- pa_only_artis %>%
    filter(presence_or_absence == 1, !is.na(order)) %>%
    mutate(country = strsplit(countries_found, " ")) %>%
    unnest(country) %>%
    distinct(order, country) %>%
    group_by(order) %>%
    summarize(countries_for_genus = paste(sort(unique(country)), collapse = " "),
              .groups = "drop") %>%
    {
      setNames(.$countries_for_genus, .$order)
    } %>%
    data.table(
      taxa_name = names(.),
      countries_found = .,
      taxa_level = "order"
    )
  
  # Class
  class_country_lookup <- pa_only_artis %>%
    filter(presence_or_absence == 1, !is.na(class)) %>%
    mutate(country = strsplit(countries_found, " ")) %>%
    unnest(country) %>%
    distinct(class, country) %>%
    group_by(class) %>%
    summarize(countries_for_genus = paste(sort(unique(country)), collapse = " "),
              .groups = "drop") %>%
    {
      setNames(.$countries_for_genus, .$class)
    } %>%
    data.table(
      taxa_name = names(.),
      countries_found = .,
      taxa_level = "class"
    )
  
  # Superclass
  superclass_country_lookup <- pa_only_artis %>%
    filter(presence_or_absence == 1, !is.na(superclass)) %>%
    mutate(country = strsplit(countries_found, " ")) %>%
    unnest(country) %>%
    distinct(superclass, country) %>%
    group_by(superclass) %>%
    summarize(countries_for_genus = paste(sort(unique(country)), collapse = " "),
              .groups = "drop") %>%
    {
      setNames(.$countries_for_genus, .$superclass)
    } %>%
    data.table(
      taxa_name = names(.),
      countries_found = .,
      taxa_level = "superclass"
    )
  
  # Phylum
  phylum_country_lookup <- pa_only_artis %>%
    filter(presence_or_absence == 1, !is.na(phylum)) %>%
    mutate(country = strsplit(countries_found, " ")) %>%
    unnest(country) %>%
    distinct(phylum, country) %>%
    group_by(phylum) %>%
    summarize(countries_for_genus = paste(sort(unique(country)), collapse = " "),
              .groups = "drop") %>%
    {
      setNames(.$countries_for_genus, .$phylum)
    } %>%
    data.table(
      taxa_name = names(.),
      countries_found = .,
      taxa_level = "phylum"
    )
  
  # Final data to be printed
  data.table(
    bind_rows(
      genus_country_lookup,
      subfamily_country_lookup,
      family_country_lookup,
      order_country_lookup,
      class_country_lookup,
      superclass_country_lookup,
      phylum_country_lookup
    )
  ) %>%
    select(-taxa_level)
  
}
```

```{r consuption data reading}
# Read in consumption data
consumption_all_years <- read_parquet("../data/consumption/example_consumption_eez_2024_12_06.parquet")

# Filter working data to just 2019
consumption <- consumption_all_years %>%
  filter(year == 2019)

# Read in taxonomic data
sciname <- fread("../data/consumption/sciname.csv")

# Get unique taxa names by eez (excluding trade flows)
scinames_excluding_tradeflows <- consumption_all_years %>%
  ungroup() %>%
  distinct(eez_iso3c, sciname_hs_modified) %>%
  left_join(sciname, by = c("sciname_hs_modified" = "sciname")) %>%
  rename(sciname = "sciname_hs_modified") %>%
  taxa_level_vars() %>%
  rename(sciname_hs_modified = "sciname") %>%
  filter(!is.na(sciname_hs_modified)) %>%
  ungroup()

# Write to parquet
write_parquet(scinames_excluding_tradeflows, "../output/scinames_excluding_tradeflows.parquet")
```

```{r get fb/slb data}
if (!file.exists("../data/fb_slb_data/")) {
  
  # Create new directory for FB/SLB data
  dir.create("../data/fb_slb_data")
  
  # Obtain FB / SLB country() data
  fb_presences_country <- country(server = "fishbase")
  slb_presences_country <- country(server = "sealifebase")
  
  fb_species_codes <- load_taxa(server = "fishbase") %>%
    rename(ScientificName = "Species") %>%
    mutate(Species = sapply(str_split(ScientificName, " "), `[`, 2)) %>%
    relocate(Species, .before = Genus)

  slb_species_codes <- load_taxa(server = "sealifebase") %>%
    rename(ScientificName = "Species") %>%
    mutate(Species = sapply(str_split(ScientificName, " "), `[`, 2)) %>%
    relocate(Species, .before = Genus)
  
  # Join FB / SLB country() to species codes
  fb_presences_country <- fb_presences_country %>%
    left_join(fb_species_codes, by = "SpecCode") %>%
    rename(SubFamily = "Subfamily") %>%
    mutate(Phylum = NA_character_, Kingdom = NA_character_) %>% # Create new 
    relocate(c(ScientificName, Species, Genus, SubFamily, Family, Order, Class, SuperClass, Phylum, Kingdom), .after = SpecCode) %>% # Move scinames to the front of dataset
    mutate(across(c(ScientificName, Species, Genus, SubFamily, Family, Order, Class, SuperClass, Phylum, Kingdom), ~ str_to_lower(.))) # convert scinames to lowercase
  
  # Same thing but for sealifebase
  slb_presences_country <- slb_presences_country %>%
    left_join(slb_species_codes, by = "SpecCode") %>%
    rename(SubFamily = "Subfamily") %>%
    mutate(SuperClass = NA_character_) %>%
    relocate(c(ScientificName, Species, Genus, SubFamily, Family, Order, Class, SuperClass, Phylum, Kingdom), .after = SpecCode) %>%
    mutate(across(c(ScientificName, Species, Genus, SubFamily, Family, Order, Class, SuperClass, Phylum, Kingdom), ~ str_to_lower(.))) 
  
  # Write files to .parquet
  write_parquet(fb_presences_country, "../data/fb_slb_data/fb_presences_country.parquet")
  write_parquet(slb_presences_country, "../data/fb_slb_data/slb_presences_country.parquet")
  write_parquet(slb_species_codes, "../data/fb_slb_data/fb_species_codes.parquet")
  write_parquet(slb_species_codes, "../data/fb_slb_data/slb_species_codes.parquet")
}
```

```{r add in fishbase / sealifebase data connections}
# Read in consumption data
scinames_excluding_tradeflows <- read_parquet("../output/scinames_excluding_tradeflows.parquet")

# Read in fishbase/sealifebase data
fb_presences_country <- read_parquet("../data/fb_slb_data/fb_presences_country.parquet")
slb_presences_country <- read_parquet("../data/fb_slb_data/slb_presences_country.parquet")

fb_species_codes <- read_parquet("../data/fb_slb_data/fb_species_codes.parquet")
slb_species_codes <- read_parquet("../data/fb_slb_data/slb_species_codes.parquet")

# Read in unaligned species names
unaligned_names <- fread("../data/fb_slb_data/unaligned_names.csv")

# Look at which artis species names differ from fb/slb species names (ARTIS is more up to date)
fb_slb <- bind_rows(fb_presences_country, slb_presences_country)

setdiff(sciname$sciname[str_detect(sciname$sciname, " ")],fb_slb$ScientificName) 

scinames_excluding_tradeflows <- scinames_excluding_tradeflows %>%
  mutate(sciname_joinable_to_fishbase = sciname_hs_modified)

# ***
# MAKING SURE ARTIS SCINAMES ALIGN WITH FISHBASE

# # of ARTIS unique scientific names (includes higher taxonomic levels that aren't only species) - 1771 total scientific names
length(unique(scinames_excluding_tradeflows$sciname_joinable_to_fishbase))

# # of ARTIS species (1421 species)
scinames_excluding_tradeflows %>%
  ungroup() %>%
  filter(str_detect(sciname_hs_modified, " ")) %>%
  distinct(sciname_hs_modified) %>%
  nrow()

# METHOD 2: rfishbase::country()

# Obtain stats on fishbase/sealifebase data
taxa_presence_or_absence(data = "fishbase", obtain_stats = TRUE)
taxa_presence_or_absence(data = "sealifebase", obtain_stats = TRUE)

# Create presence/absence data
fishbase_presences_country <- taxa_presence_or_absence(data = "fishbase")
sealifebase_presences_country <- taxa_presence_or_absence(data = "sealifebase")

# Number of rows in ARTIS data (fishbase + sealifebase presence data)
nrow(fishbase_presences_country)
nrow(sealifebase_presences_country)

# Combine fishbase/sealifebase data
merged_presence_absence_data_country <- bind_rows(fishbase_presences_country[-which(sealifebase_presences_country$source == "sealifebase"),], sealifebase_presences_country %>% filter(source == "sealifebase"))

# Add true species column
merged_presence_absence_data_country <- merged_presence_absence_data_country %>%
  mutate(species = case_when(str_detect(sciname_hs_modified, " ") ~ sapply(str_split(sciname_hs_modified, " "), `[`, 2),
                             TRUE ~ NA))

# Determine if lower resolution taxa are in an eez (e.g., if a species for a given country is found there that is also a lower resolution taxa like actinoptergyii, then all actinoptergyii or higher will be found in that country) -

# Create a dataset to identify whether a species comes from ARTIS or slb/fb
# List of ARTIS species that aren't matched to either fishbase or sealifebase - this is what can be derived from this code
species_determination <- full_join(
  bind_rows(fb_presences_country %>% 
              mutate(source = "fishbase"), 
            slb_presences_country %>% 
              mutate(source = "sealifebase")) %>%
    mutate(ScientificName = tolower(ScientificName)) %>%
    select(ScientificName) %>%
    distinct() %>% 
    mutate(fb_slb = 1), 
  sciname %>% 
    select(sciname, common_name) %>%
    filter(str_detect(sciname, " ")) %>% 
    distinct() %>% 
    mutate(artis = 1), 
  by = c("ScientificName" = "sciname")) 

# Species in ARTIS but not in FB/SLB
artis_species_not_in_fbslb <- species_determination %>%
  filter(is.na(fb_slb) == TRUE, artis == 1) %>%
  pull(ScientificName)

# Species in FB/SLB but not ARTIS
fbslb_species_not_in_artis <- species_determination %>%
  filter(fb_slb == 1, is.na(artis) == TRUE) %>%
  pull(ScientificName)

# Prepare data to increase presence / absence for all taxa, not just true species
pa_only_artis <- merged_presence_absence_data_country %>%
  mutate(presence_or_absence = case_when(
    sciname_joinable_to_fishbase %in% fbslb_species_not_in_artis ~ NA, 
    TRUE ~ presence_or_absence))

# Get distributions that taxa are present in for a given country
country_taxa_lookups <- get_country_taxa_data(pa_only_artis)

# Get presences and absences for taxa levels higher than true species
presence_absence_data <- pa_only_artis %>%
  left_join(country_taxa_lookups,
            by = c("sciname_hs_modified" = "taxa_name")) %>%
  mutate(
    countries_found = case_when(
      !countries_found.x == "" ~ countries_found.x,!countries_found.y == "" ~ countries_found.y,
      TRUE ~ NA_character_
    )
  ) %>%
  mutate(presence_or_absence_all_taxa = case_when(str_detect(countries_found, eez_iso3c) ~ 1, TRUE ~ 0)) %>%
  select(-countries_found.x, -countries_found.y, -presence_or_absence,
         -sciname_joinable_to_fishbase, -new_sciname, -joined, -source) %>%
  relocate(species, .before = genus)

  # Remove extraneous objects to save memory to make the following left_join faster
rm(consumption, fb_presences_country, fb_slb, fb_species_codes, fishbase_presences_country,
   merged_presence_absence_data_country, pa_only_artis, sciname, sciname_metadata,
   scinames_excluding_tradeflows, sealifebase_presences_country, slb_presences_country,
   slb_species_codes, species_determination, unaligned_names, artis_species_not_in_fbslb,
   fbslb_species_not_in_artis, country_taxa_lookups)

# Garbage collection
gc()

# Join to consumption data
consumption_presence_absence <- left_join(consumption_all_years, presence_absence_data, by = c("sciname_hs_modified", "eez_iso3c")) %>%
  relocate(eez_iso3c, .after = "consumer_iso3c")

# Write to .parquet
write_parquet(presence_absence_data, "../output/presence_absence_data.parquet")

# Write to .parquet
write_parquet(consumption_presence_absence, "../output/consumption_presence_absence.parquet")

# Remove more objects
rm(consumption_all_years, presence_absence_data)

# Garbage collection
gc()
```
