==> code/01_Taxa_Disaggregation.Rmd <==
---
title: "Taxa disaggregation"
author: "Connor Quiroz"
date: "2025-06-22"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
# General data cleaning + visualization + analysis packages
library(tidyverse)
library(ggformula)
library(countrycode)
library(data.table)
library(arrow)
library(cowplot)
library(RColorBrewer)

# Map creation
library(rnaturalearth)
library(rnaturalearthdata)
library(sf)

# Load in packages needed for ARTIS
library(devtools)
library(tidytext)
# devtools::install_github("davidsjoberg/ggsankey")
library(ggsankey)
# devtools::install_github("Seafood-Globalization-Lab/exploreARTIS@v1.0.0", dependencies = TRUE)
library(exploreARTIS)

# For Packages for fishbase
# remotes::install_github("cboettig/duckdbfs", force = TRUE)
# remotes::install_github("ropensci/rfishbase")
library(duckdbfs)
library(rfishbase)
```

```{r initialize functions}
# Ocean polygon
ocean <- st_polygon(list(cbind(c(seq(-180, 179, len = 100), rep(180, 100), 
                        seq(179, -180, len = 100), rep(-180, 100)),
                      c(rep(-90, 100), seq(-89, 89, len = 100),
                        rep(90, 100), seq(89, -90, len = 100))))) %>%
  st_sfc(crs = "WGS84") %>%
  st_as_sf()

# Map creation function
create_map <- function(data = data, fill = "prop_missing", country.col.name) {
  world_map <- ne_countries(scale = "medium", returnclass = "sf")

world_map <- world_map %>%
  left_join(data, by = c("iso_a3" = country.col.name))
  
  world_map %>%
ggplot() +
  geom_sf(data = ocean, fill = "#8080ff80") +
  geom_sf(aes(fill = !!sym(fill)), color = "black") + 
  scale_fill_viridis_c(option = "plasma") +
  theme_minimal() +
  theme(legend.position = "bottom") +
  coord_sf(crs = "+proj=robin")
}
```


```{r preprocess consumption data to add sciname classification}
# Read in prsence absence consumption data
consumption_presence_absence <- read_parquet("../output/consumption_presence_absence.parquet")

# Read in sciname data
sciname <- fread("../data/consumption/sciname.csv") 

consumption_presence_absence <- left_join(consumption_presence_absence, 
                                          sciname, 
                                          by = c("sciname_hs_modified" = "sciname")) %>%
  taxa_level_vars(taxa_var = "sciname_hs_modified")

# Garbage collecter to clean up storage
gc()

consumption_presence_absence %>%
  group_by(eez_iso3c) %>%
  drop_na(presence_or_absence_all_taxa) %>%
  summarize(sum_presences = sum(presence_or_absence_all_taxa)) %>%
  create_map(fill = "sum_presences", country.col.name = "eez_iso3c")

# Group by eez
consumption_presence_absence %>%
  group_by(eez_iso3c) %>%
  summarize(prop_presence = mean(presence_or_absence_all_taxa, na.rm = TRUE)) %>%
  create_map(fill = "prop_presence", country.col.name = "eez_iso3c")
  
```

```{r explore dataset problems}
scinames_excluding_tradeflows <- read_parquet("../output/scinames_excluding_tradeflows.parquet")

# Scenario 1
taxa_levels <- c("kingdom","phylum","superclass","class", "order", "family", "genus", "species")
scenario_1 <- c()
for (i in 1:(length(taxa_levels) - 1)) {
  value <- scinames_excluding_tradeflows %>%
    mutate(
      species = case_when(
        str_count(sciname_hs_modified, " ") == 1 ~ str_split(sciname_hs_modified, " ") %>% 
          sapply(function(x) x[2]), # Extract second word if exactly two words
        TRUE ~ NA_character_ # Return NA if not two words
      )
    ) %>%
    group_by_at(taxa_levels[i]) %>%  # Group by the current taxa level (e.g., class)
    filter(!is.na(get(taxa_levels[i]))) %>%  # Filter out NA values for the current level
    summarize(
      all_next_na = all(!is.na(get(taxa_levels[i + 1]))),  # Check if the next taxa level is NA
      .groups = "drop"
    ) %>%
    summarize(sum = sum(all_next_na)) %>% # Summarize how many rows have all NA for the next taxa level
    pull(sum)
  scenario_1 <- append(scenario_1, value)
}

####################
####################
####################


# Scenario 2

scenario_2 <- c()
for (i in 1:(length(taxa_levels) - 1)) {
  value <- scinames_excluding_tradeflows %>%
    mutate(
      species = case_when(
        str_count(sciname_hs_modified, " ") == 1 ~ str_split(sciname_hs_modified, " ") %>% 
          sapply(function(x) x[2]), # Extract second word if exactly two words
        TRUE ~ NA_character_ # Return NA if not two words
      )
    ) %>%
    group_by_at(taxa_levels[i]) %>%  # Group by the current taxa level (e.g., class)
    filter(!is.na(get(taxa_levels[i]))) %>%  # Filter out NA values for the current level
    summarize(
      all_next_na = all(is.na(get(taxa_levels[i + 1]))),  # Check if the next taxa level is NA
      .groups = "drop"
    ) %>%
    summarize(sum = sum(all_next_na)) %>% # Summarize how many rows have all NA for the next taxa level
    pull(sum)
  scenario_2 <- append(scenario_2, value)
}


scinames_excluding_tradeflows %>%
  mutate(
    species = case_when(
      str_count(sciname_hs_modified, " ") == 1 ~ str_split(sciname_hs_modified, " ") %>%
        sapply(function(x) x[2]), # Extract second word if exactly two words
      TRUE ~ NA_character_ # Return NA if not two words
    )
  ) %>%
  group_by(order) %>%
  filter(!is.na(order)) %>%
  summarize(
    all_genus_na = all(is.na(family)),  # Check if all rows for a class have NA in genus
    .groups = "drop"
  ) %>%
  summarize(sum = sum(all_genus_na))

####################
####################
####################

# Scenario 3
scenario_3 <- c()
for (i in 1:(length(taxa_levels) - 1)) {
  value <- scinames_excluding_tradeflows %>%
    mutate(
      species = case_when(
        str_count(sciname_hs_modified, " ") == 1 ~ str_split(sciname_hs_modified, " ") %>% 
          sapply(function(x) x[2]), # Extract second word if exactly two words
        TRUE ~ NA_character_ # Return NA if not two words
      )
    ) %>%
    group_by_at(taxa_levels[i]) %>%  # Group by the current taxa level (e.g., class)
    filter(!is.na(get(taxa_levels[i]))) %>%  # Filter out NA values for the current level
    summarize(
      all_next_na = any(is.na(get(taxa_levels[i + 1]))),  # Check if the next taxa level is NA
      .groups = "drop"
    ) %>%
    summarize(sum = sum(all_next_na)) %>% # Summarize how many rows have all NA for the next taxa level
    pull(sum)
  scenario_3 <- append(scenario_3, value)
}

# Correct scenario 3 counts (scenario 3 originally includes stuff from scenario 2, so subtract if 2 > 3)
for (i in 1:4) {
  if(scenario_3[i] >= scenario_2[i]) {
    scenario_3[i] <- scenario_3[i] - scenario_2[i]
  }
}

# Create plot
data.frame(missing_taxa = c(scenario_1, scenario_2, scenario_3),
           scenario = rep(c("1", "2", "3"), each = 7),
           taxa_level = c("Kingdom","Phylum","Superclass","Class", "Order", "Family", "Genus")) %>%
  ggplot(aes(x = factor(taxa_level, levels = c("Genus", "Family", "Order", "Class", "Superclass","Phylum", "Kingdom")), 
             y = missing_taxa, 
             fill = scenario)) +
  geom_col(position = "fill") +
  labs(x = "Taxa level", y = "Missing taxa", fill = "Scenario") +
  theme_cowplot() +
  scale_fill_viridis_d(end = 0.8) +
  theme(axis.text = element_text(size = 8))
```
